loan_app/requirements.txt
Flask==2.3.3
scikit-learn==1.6.1
xgboost==2.0.0
pandas==2.1.4
numpy==1.26.4
joblib==1.4.0
python-dotenv==1.0.0
dashscope==1.20.0
google-generativeai
